Isolamento de componente, refatoração de funcionalidade de seleção de números no jogo. Análise de desempenho.
